refactor: refatorying 진행
