+++ src/components/common/notification/_components/notification-list/_components/notification-tab/_components/format-notification-meta-time.constants.ts
@@ -1,0 +1,3 @@
+export const MINUTE_MS = 60_000;
+export const HOUR_MS = 3_600_000;
+export const DAY_MS = 86_400_000;
