+++ src/components/common/notification/_components/notification-list/_components/notification-tab/notification-tab.constants.ts
@@ -1,0 +1,7 @@
+export const EMPTY_COMMENTS_FALLBACK = '클라이밍 어때요?-딸기님의 댓글"정말재밌어요:)"';
+
+export const NOTIFICATION_ROW_BUTTON_CLASS =
+  'h-auto w-full shrink-0 flex flex-col items-stretch justify-start gap-0 rounded-none border-0 p-0 text-left font-normal whitespace-normal shadow-none hover:bg-transparent';
+
+export const NOTIFICATION_ICON_BOX_CLASS =
+  'flex size-[42px] shrink-0 items-center justify-center rounded-[14px] px-[9.5px] text-white';
